Encryption and Decryption use the secret symetric key
Previous example use the public encrypted key as symetric key to encrypt and decrypt in the tests.

Now the secret symetric key is decoded with the master_key (derived from email and password). This is the key that is used to encrypt and decrypt.

It probably won't change the issue of the test itself (anything of the correct size could be used as a key) but it won't confuse the readers about the real key usage.