config.ru: set RACK_ENV to deployment if RUBYWARDEN_ENV is production
If RACK_ENV is not set, which it may not be since RUBYWARDEN_ENV now
has to be used, it will default to development, which will cause
Sinatra to print stack traces and things to the public.

Protect users by setting RACK_ENV to "deployment" when
RUBYWARDEN_ENV is set to "production".