Identity: Return user Kdf and KdfIterations in login response
It used to be enough just to return these in /accounts/prelogin but
now, without them, the browser extension stores a null KDF iteration
count upon login and then on subsequent unlock attempts, it uses the
default KDF iteration count of 5000 which is probably not right (the
new default is 100,000).  This causes it to generate a failed key,
making it impossible to unlock a logged-in vault without logging out
and logging back in.

See https://github.com/bitwarden/browser/issues/1456